+++ src/js/tests/testHaversine.js
@@ -1,5 +1,5 @@
-import Vertex from "./classes/Vertex.js";
-import Graph from "./classes/Graph.js";
+import Vertex from "../classes/Vertex.js";
+import Graph from "../classes/Graph.js";
 import { strict as assert } from 'assert';
 
 const vertex1 = new Vertex("loc1", "intersection", -7.745820063276957, 110.39869308471681, null, null, ["loc2", "loc3", "loc4", "loc5", "loc6", "loc7"]);
